Created directories for routes and controllers to clean up our server.js file. Broke up the 'GET' routes on server.js into the new directories.
