+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "playlistr",
 	"version": "1.0.0",
-	"description": "Mern Demo",
+	"description": "",
 	"main": "server.js",
 	"scripts": {
 		"start": "if-env NODE_ENV=production && npm run start:prod || npm run start:dev",
