+++ server.js
@@ -1,48 +1,13 @@
-let express = require('express');
-let request = require('request');
-let querystring = require('querystring');
+const express = require('express');
+const routes = require('./routes');
+const app = express();
+const port = process.env.PORT || 8888;
 
-require('dotenv').config();
+if (process.env.NODE_ENV === 'production') {
+	app.use(express.static('client/build'));
+}
 
-let app = express();
+app.use(routes);
 
-let redirect_uri = process.env.REDIRECT_URI || 'http://localhost:8888/callback';
-
-app.get('/spotify/login', function (req, res) {
-	res.redirect(
-		'https://accounts.spotify.com/authorize?' +
-			querystring.stringify({
-				response_type: 'code',
-				client_id: process.env.client_id,
-				scope: 'user-read-private user-read-email',
-				redirect_uri
-			})
-	);
-});
-
-app.get('/callback', function (req, res) {
-	let code = req.query.code || null;
-	let authOptions = {
-		url: 'https://accounts.spotify.com/api/token',
-		form: {
-			code: code,
-			redirect_uri,
-			grant_type: 'authorization_code'
-		},
-		headers: {
-			Authorization:
-				'Basic ' +
-				Buffer.from(process.env.client_id + ':' + process.env.client_secret).toString('base64')
-		},
-		json: true
-	};
-	request.post(authOptions, function (error, response, body) {
-		var access_token = body.access_token;
-		let uri = process.env.FRONTEND_URI || 'http://localhost:3000/home';
-		res.redirect(uri + '?access_token=' + access_token);
-	});
-});
-
-let port = process.env.PORT || 8888;
-console.log(`Listening on port ${port}. Go /login to initiate authentication flow.`);
+console.log(`Listening on port ${port}.`);
 app.listen(port);
